Added guidance for ASN number
Currently there is a bug in Azure where we allow the user to select an invalid ASN number. If they do so, there is no error, and later they can't create BGP connections (with no error visible to them about what is wrong).

Hopefully this FAQ will save Azure users some pain.

I plan to also submit an update to the PowerShell cmd-lets to add a check for a valid ASN number.